requirements.txt
# ── Framework principal ───────────────────────────────────────────────────────
Scrapy>=2.11.0
scrapy-playwright>=0.0.33

# ── CLI ───────────────────────────────────────────────────────────────────────
click>=8.1.7

# ── Parsing & extraction ──────────────────────────────────────────────────────
itemadapter>=0.9.0
lxml>=5.1.0
cssselect>=1.2.0

# ── Proxies & YAML ────────────────────────────────────────────────────────────
PyYAML>=6.0.1

# ── Tests ─────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
scrapy-fake-useragent>=1.4.4

# ── Utilitaires ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
